pyproject.toml
[tool.poetry]
name = "vertex-ai-pipeline-example"
version = "0.1.0"
description = ""
authors = ["fittecs <jun.fujita@fittecs.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11.0, <3.12"
kfp = "2.7.0"
google-cloud-pipeline-components = "2.17.0"
google-cloud-aiplatform = "1.71.1"


[tool.poetry.group.dev.dependencies]
pytest = "8.3.3"
pytest-mock = "3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
